Rename the namespace of HttpRequestMessagesCheckExtensions tests to actually contain the name HttpRequestMessagesCheckExtensions.
