+++ test/TestableHttpClient.Tests/HttpRequestMessagesCheckExtensionsTests/WithContentHeaderNameAndValue.cs
@@ -4,7 +4,7 @@ using Moq;
 
 using Xunit;
 
-namespace TestableHttpClient.Tests.HttpRequestMessagesExtensionsTests
+namespace TestableHttpClient.Tests.HttpRequestMessagesCheckExtensionsTests
 {
     public class WithContentHeaderNameAndValue
     {
