+++ test/TestableHttpClient.Tests/HttpRequestMessagesCheckExtensionsTests/WithHttpMethod.cs
@@ -5,7 +5,7 @@ using Moq;
 
 using Xunit;
 
-namespace TestableHttpClient.Tests.HttpRequestMessagesExtensionsTests
+namespace TestableHttpClient.Tests.HttpRequestMessagesCheckExtensionsTests
 {
     public class WithHttpMethod
     {
